Refactor filter logic
